WIP: Show example of error handling for readJson
RE: Fix Regression in PATCH on system node

As part of that discussion on error handling, it seems that there were
some missteps in readjson that need corrected.  This is my proposal for
how to correct them.

Tested:
Not yet tested.  Example only.

+++ include/persistent_data_middleware.hpp
@@ -24,7 +24,7 @@ class Middleware
 {
     // todo(ed) should read this from a fixed location somewhere, not CWD
     static constexpr const char* filename = "bmcweb_persistent_data.json";
-    int jsonRevision = 1;
+    uint64_t jsonRevision = 1;
 
   public:
     struct Context
@@ -58,7 +58,7 @@ class Middleware
     void readData()
     {
         std::ifstream persistentFile(filename);
-        int fileRevision = 0;
+        uint64_t fileRevision = 0;
         if (persistentFile.is_open())
         {
             // call with exceptions disabled
